+++ include/solvers/rk_solver.hpp
@@ -2,12 +2,12 @@
 #include <functional>
 
 namespace solvers {
-using Function = std::function<double(double, double)>;
+using Function = std::function<float(float, float)>;
 class Solver {
     public:
-	Solver(double x0, double y0, double x, double h, const Function func);
-	double x0, y0, x, h, xn, yn;
-	virtual double solve() = 0;
+	Solver(float x0, float y0, float x, float h, const Function func);
+	float x0, y0, x, h, xn, yn;
+	virtual float solve() = 0;
 	const Function func;
 
 };
@@ -16,7 +16,7 @@ namespace euler { //basically rk2
 class EulerSolver : public Solver {
     public:
 	using Solver::Solver;	
-	double solve() override;
+	float solve() override;
 
 };
 }
@@ -25,7 +25,7 @@ namespace rk4 {
 class RK4Solver : public Solver {
     public:
 	using Solver::Solver;	
-	double solve() override;
+	float solve() override;
 
 };
 }
