+++ src/solvers/rk4.cpp
@@ -4,12 +4,12 @@
 namespace solvers {
 namespace rk4{
 	
-double RK4Solver::solve(){
+float RK4Solver::solve(){
     while(xn < x){
-	double k1 = func(xn, yn);
-	double k2 = func(xn + (h/2), (yn + ((h/2) * k1)));
-	double k3 = func(xn + (h/2), (yn + ((h/2) * k2)));
-	double k4 = func(xn + h, yn + (h * k3));
+	float k1 = func(xn, yn);
+	float k2 = func(xn + (h/2), (yn + ((h/2) * k1)));
+	float k3 = func(xn + (h/2), (yn + ((h/2) * k2)));
+	float k4 = func(xn + h, yn + (h * k3));
 	yn += ((h/6)*(k1 + (2*k2) + (2*k3) + k4));
 	xn += h;
     }
