+++ src/solvers/solver.cpp
@@ -3,6 +3,6 @@
 
 
 namespace solvers {
-    Solver::Solver(double x0, double y0, double x, double h, const Function func) : x0(x0), y0(y0), x(x), h(h), func(func), xn(x0), yn(y0){};
+    Solver::Solver(float x0, float y0, float x, float h, const Function func) : x0(x0), y0(y0), x(x), h(h), func(func), xn(x0), yn(y0){};
 }
